App_TTS/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-aarch64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=51_gnu
anyio=4.3.0=pypi_0
archspec=0.2.1=pyhd3eb1b0_0
boltons=23.0.0=py39hd43f75c_0
brotli-python=1.0.9=py39h419075a_7
bzip2=1.0.8=hfd63f10_2
c-ares=1.19.1=h998d150_0
ca-certificates=2023.12.12=hd43f75c_0
certifi=2024.2.2=py39hd43f75c_0
cffi=1.16.0=py39h998d150_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
coloredlogs=15.0.1=pypi_0
conda=24.1.2=py39hd43f75c_0
conda-content-trust=0.2.0=py39hd43f75c_0
conda-libmamba-solver=23.12.0=pyhd3eb1b0_1
conda-package-handling=2.2.0=py39hd43f75c_0
conda-package-streaming=0.9.0=py39hd43f75c_0
cryptography=42.0.2=py39h5077475_0
dimits=0.0.21a0=pypi_0
distro=1.8.0=py39hd43f75c_0
espeak-phonemizer=1.3.1=pypi_0
exceptiongroup=1.2.0=pypi_0
expat=2.5.0=h419075a_0
flatbuffers=23.5.26=pypi_0
fmt=9.1.0=hb8fdbf2_0
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.25.2=pypi_0
humanfriendly=10.0=pypi_0
icu=73.1=h419075a_0
idna=3.4=py39hd43f75c_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
krb5=1.20.1=h2e2fba8_1
ld_impl_linux-aarch64=2.38=h8131f2d_1
libarchive=3.6.2=h654c02d_2
libcurl=8.5.0=hfa2bbb0_0
libedit=3.1.20230828=h998d150_0
libev=4.33=hfd63f10_1
libffi=3.4.4=h419075a_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libmamba=1.5.3=h78dbd8a_0
libmambapy=1.5.3=py39hd82f176_0
libnghttp2=1.57.0=hb788212_0
libsolv=0.7.24=h94b7715_0
libssh2=1.10.0=h6ac735f_2
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h998d150_0
libxml2=2.10.4=h045d036_1
lz4-c=1.9.4=h419075a_0
menuinst=2.0.2=py39hd43f75c_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h419075a_0
numpy=1.26.4=pypi_0
ollama=0.1.6=pypi_0
onnxruntime=1.17.1=pypi_0
openssl=3.0.13=h2f4d8fa_0
packaging=23.1=py39hd43f75c_0
pcre2=10.42=hcfaa891_0
pip=23.3.1=py39hd43f75c_0
piper-phonemize=1.1.0=pypi_0
piper-tts=1.2.0=pypi_0
platformdirs=3.10.0=py39hd43f75c_0
pluggy=1.0.0=py39hd43f75c_1
protobuf=4.25.3=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycosat=0.6.6=py39h998d150_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.5.2=pypi_0
pysocks=1.7.1=py39hd43f75c_0
python=3.9.18=h4bb2201_0
readline=8.2=h998d150_0
reproc=14.2.4=h22f4aa5_1
reproc-cpp=14.2.4=h22f4aa5_1
requests=2.31.0=py39hd43f75c_1
ruamel.yaml=0.17.21=py39h998d150_0
ruamel.yaml.clib=0.2.6=py39h998d150_1
setuptools=68.2.2=py39hd43f75c_0
sniffio=1.3.1=pypi_0
sqlite=3.41.2=h998d150_0
sympy=1.12=pypi_0
tk=8.6.12=h241ca14_0
tqdm=4.64.0=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2023d=h04d1e81_0
urllib3=2.1.0=py39hd43f75c_1
wheel=0.41.2=py39hd43f75c_0
xz=5.4.5=h998d150_0
yaml-cpp=0.8.0=h419075a_0
zlib=1.2.13=h998d150_0
zstandard=0.19.0=py39h998d150_0
zstd=1.5.5=h6a09583_0
